Move Configuration to top-level as APIServiceConfiguration
Deprecate abstract superclass APIService in favor of protocol APIService
