+++ OBAKitCore/Network/APIService.swift
@@ -6,21 +6,67 @@
 //  This source code is licensed under the Apache 2.0 license found in the
 //  LICENSE file in the root directory of this source tree.
 //
+public protocol APIService {
+    var configuration: APIServiceConfiguration { get }
+    var dataLoader: URLDataLoader { get }
 
-import Foundation
+    /// - parameter dataLoader: For mocking network requests. Use `URLSession.shared` for production cases.
+    init(_ configuration: APIServiceConfiguration, dataLoader: URLDataLoader)
+}
+
+public struct APIServiceConfiguration {
+    let baseURL: URL
+    let apiKey: String
+    let uuid: String
+    let appVersion: String
+    let regionIdentifier: Int?
+
+    /// Generated query items from the initializer. Include this set of query items with every request you make.
+    let defaultQueryItems: [URLQueryItem]
+
+    /// - parameter baseURL: Any queries included in the `baseURL` will be moved to the `defaultQueryItems`.
+    init(baseURL: URL, apiKey: String, uuid: String, appVersion: String, regionIdentifier: Int?) {
+        self.apiKey = apiKey
+        self.uuid = uuid
+        self.appVersion = appVersion
+        self.regionIdentifier = regionIdentifier
+
+        // Add default query items, including the component query of the baseURL.
+        var queryItems: [URLQueryItem] = [
+            "key": apiKey,
+            "app_uid": uuid,
+            "app_ver": appVersion,
+            "version": "2"
+        ].map(URLQueryItem.init(name:value:))
+
+        var components = URLComponents(url: baseURL, resolvingAgainstBaseURL: false)!
+        if let items = components.queryItems, items.isEmpty == false {
+            queryItems.append(contentsOf: items)
+        }
+
+        self.defaultQueryItems = queryItems
+
+        // We've successfully saved off any query items that were a part of the baseURL.
+        // So we nil out the query here so that we can safely append path components to
+        // the baseURL elsewhere.
+        components.query = nil
+
+        self.baseURL = components.url!
+    }
+}
 
 /// The core class for interacting with the OBA REST API. Treat this as an abstract class, and instead interact with its subclasses.
 ///
 /// This class is responsible for managing the base URL for the OBA REST API server,
 /// the network queue for in-flight operations, and the query parameters that are
 /// common to every request.
-public class APIService: NSObject {
+public class _APIService: NSObject {
     let baseURL: URL
     let networkQueue: OperationQueue
     let defaultQueryItems: [URLQueryItem]
     let dataLoader: URLDataLoader
 
-    /// Creates a new instance of `APIService`.
+    /// Creates a new instance of `_APIService`.
     /// - Parameters:
     ///   - baseURL: The base URL for the service you will be using.
     ///   - apiKey: The API key for the service you will be using. Passed along as `key` in query params.
@@ -56,7 +102,7 @@ public class APIService: NSObject {
         self.dataLoader = dataLoader
     }
 
-    /// Creates a new instance of APIService.
+    /// Creates a new instance of _APIService.
     /// - Parameters:
     ///   - baseURL: The base URL for the service you will be using.
     ///   - apiKey: The API key for the service you will be using. Passed along as `key` in query params.
